Add build version tooling for CircleCI
This adds centrally maintained parametrized versioning for the building
of docker images. For the three built images:
- db
- synapse
- well_known_server
the software versions that are supposed to land in the docker images are
defined in the file `BUILD_VERSIONS`.
The circleCI configuration sources these versions in the `.bashrc`
envirnoment and passes corresponding `--build-arg`s during the build
stage.
The `Dockerfile`s now expect the build args to be defined - or will fail
to build.