Add runtime reloading of the federation whitelist
Previously the 'purger' service would periodically check if the
configured whitelist of known federation servers was still up to date
and if not restart the affected synapse services.

That caused downtime and also was a somewhat annoying workaround.

This now introduces a module that checks and updates the configuration
at runtime without the need for a restart.

Note that this is accessing Synapse internals that aren't considered
stable and therefore subject to change with future version upgrades.