fix Synapse configuration file format
For unknown reasons the Synapse module by Raiden for the presence started to
throw issues on parsing the configuration file. The `service_registry_address`
option in the module configuration gets parsed as an integer. In the following
the `eth_utils` functions fail to parse this number as a valid Ethereum address.
In result the whole server fails to start.
The origin of this new problem is not known. Maybe the parsing of hex values in
a YAML file get now specially parsed and not taken are pure strings anymore. In
any case a simple solution is to explicitly make the value a string by putting
quotation marks around it in the configuration template file. This change is not
invasive and is also backwards compatible.