remove raiden dependency in synapse image
The `raiden` package and thereby also the Docker image will become depracted.
Therefore all dependent components try to get rid of it where possible. As here
it only needs to access a data value, it is quite simple to fast forward this
information into this repository here. This removes the dependency without more
work on updates or even less.