update dependencies for services and synapse
Related to the big new `raiden-contracts` release `0.50.0`, it is necessary to
update the `raiden-services` and the `raiden-synapse-modules` to stay
compatible.

Please note that the `RAIDEN_VERSION` got removed, because it is unused (only
references by the also unused `purger`). As we are currently in a transitioning
phase to depract the `raiden` package, there is no official PyPI release
anymore. Thereby it is hard to represent the current state for this version. But
as it is not used anyway, it has been decided to remove it completely to avoid
any confusion.