add new known-servers file for arbitrum one
In the past, production was a synonym for mainnet and development Goerli. As we
will now also have a federation for Arbitrum One, this must be resolved.
Therefore it has been decided to use the network names instead. Furthermore we
used the version number of the `raiden` package as postfix. As this does not
work anymore, it has been decided to use the `raiden-service-bundle` version.

The resolution for these new files does not yet work out-of-the-box and requires
RSB providers to manually define the `URL_KNOWN_FEDERATION_SERVERS` environment
variable